requirements.txt
tensorflow
librosa
streamlit
matplotlib
numpy
scikit-learn
Pillow
seaborn
soundfile
